Refactor execution/mod.rs (code motion)
