+++ src/wasm-lib/kcl/src/wasm/mod.rs
@@ -1,4 +1,4 @@
-///! Web assembly utils.
+//! Web assembly utils.
 use std::{
     pin::Pin,
     task::{Context, Poll},
